+++ tests/phpunit/tests/random-file.php
@@ -25,7 +25,7 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( ABSPATH . $random_file );
-		$this->assertRegExp( '~wp-includes/~', $random_file );
+		$this->assertMatchesRegularExpression( '~wp-includes/~', $random_file );
 	}
 
 	public function test_valid_directory_with_preceding_forward_slash() {
@@ -33,7 +33,7 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( ABSPATH . $random_file );
-		$this->assertRegExp( '~wp-includes/~', $random_file );
+		$this->assertMatchesRegularExpression( '~wp-includes/~', $random_file );
 	}
 
 	public function test_valid_directory_with_trailing_slash() {
@@ -41,7 +41,7 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( ABSPATH . $random_file );
-		$this->assertRegExp( '~wp-includes/images/~', $random_file );
+		$this->assertMatchesRegularExpression( '~wp-includes/images/~', $random_file );
 	}
 
 	public function test_valid_directory_with_subdirectory() {
@@ -49,7 +49,7 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( ABSPATH . $random_file );
-		$this->assertRegExp( '~wp-includes/images/~', $random_file );
+		$this->assertMatchesRegularExpression( '~wp-includes/images/~', $random_file );
 	}
 
 	public function test_invalid_directory() {
@@ -61,7 +61,7 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( ABSPATH . $random_file );
-		$this->assertRegExp( '~wp-includes/images/.+\.gif$~', $random_file );
+		$this->assertMatchesRegularExpression( '~wp-includes/images/.+\.gif$~', $random_file );
 	}
 
 	public function test_no_matching_extension_with_nonextistent_extension() {
@@ -95,7 +95,7 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( ABSPATH . $random_file );
-		$this->assertRegExp( '~wp-includes/images/.+\.gif$~', $random_file );
+		$this->assertMatchesRegularExpression( '~wp-includes/images/.+\.gif$~', $random_file );
 	}
 
 	public function test_matching_multiple_extensions() {
@@ -103,7 +103,7 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( ABSPATH . $random_file );
-		$this->assertRegExp( '~wp-includes/images/.+\.(jpg|gif)$~', $random_file );
+		$this->assertMatchesRegularExpression( '~wp-includes/images/.+\.(jpg|gif)$~', $random_file );
 	}
 
 	public function test_matching_multiple_extensions_with_leading_period() {
@@ -111,7 +111,7 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( ABSPATH . $random_file );
-		$this->assertRegExp( '~wp-includes/images/.+\.(jpg|gif)$~', $random_file );
+		$this->assertMatchesRegularExpression( '~wp-includes/images/.+\.(jpg|gif)$~', $random_file );
 	}
 
 	public function test_matching_multiple_extensions_via_array() {
@@ -119,7 +119,7 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( ABSPATH . $random_file );
-		$this->assertRegExp( '~wp-includes/images/.+\.(jpg|gif)$~', $random_file );
+		$this->assertMatchesRegularExpression( '~wp-includes/images/.+\.(jpg|gif)$~', $random_file );
 	}
 
 	public function test_matching_multiple_extensions_with_leading_period_via_array() {
@@ -127,7 +127,7 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( ABSPATH . $random_file );
-		$this->assertRegExp( '~wp-includes/images/.+\.(jpg|gif)$~', $random_file );
+		$this->assertMatchesRegularExpression( '~wp-includes/images/.+\.(jpg|gif)$~', $random_file );
 	}
 
 	public function test_no_matching_extension() {
@@ -143,7 +143,7 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( ABSPATH . $random_file );
-		$this->assertRegExp( '~wp-includes/images/~', $random_file );
+		$this->assertMatchesRegularExpression( '~wp-includes/images/~', $random_file );
 	}
 
 	public function test_reftype_absolute( $random_file = '' ) {
@@ -168,14 +168,14 @@ class Random_File_Test extends WP_UnitTestCase {
 
 		$this->assertNotEmpty( $random_file );
 		$this->assertFileExists( str_replace( 'http://example.org/', ABSPATH, $random_file ) );
-		$this->assertRegExp( '~^http://example.org/wp-includes/images/~', $random_file );
+		$this->assertMatchesRegularExpression( '~^http://example.org/wp-includes/images/~', $random_file );
 	}
 
 	public function test_reftype_hyperlink() {
 		$random_file = c2c_random_file( 'wp-includes/images', '', 'hyperlink' );
 
 		$this->assertNotEmpty( $random_file );
-		$this->assertRegExp( '~<a href="http://example.org/wp-includes/images/~', $random_file );
+		$this->assertMatchesRegularExpression( '~<a href="http://example.org/wp-includes/images/~', $random_file );
 	}
 
 	public function test_file_exclusion() {