+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Change: Note compatibility through WP 6.8+
 * Change: Update copyright date (2025)
 * Unit tests:
+    * Change: Switch uses of `assertRegExp()` to `assertMatchesRegularExpression()` for future PHPUnit 10 compatibility
     * Change: Explicitly define return type for overridden method
 
 ## 2.0 _(2024-08-13)_